requirements.txt
Flask-Caching==1.9.0
flask-mongoengine==0.9.5
Flask-RESTful==0.3.8
redis==3.5.3
